Add simple node script to ensure snapshot file is sorted
This is a little unfortunate, I'd like to at least do this in lua (it
should be possible?) but without **something** the snapshot file output
is unstable because packer runs async the keys of the various plugins
are often in a different order even if the revisions are the same.

Without this change, the plugins-dev.json file will change quite often
(and be very hard to actually read in `git diff` or on GitHub) even when
there are **no** changes.